+++ CHANGELOG.md
@@ -14,6 +14,10 @@ New devices:
 - Innr E14 bulb RGBW
 - Konke Multi-function button
 - Konke Temperature and humidity sensor
+- Gledopto Zigbee USB mini LED controller RGB + CCT
+- IKEA TRADFRI LED bulb E14 600 lumen, dimmable, white spectrum, opal white
+- Lingan Zigbee OnOff Controller
+- Livolo Switch TI0001
 
 ## 0.0.19
 - Add temperature sensor support for Xiaomi Mi power plug